requirements.txt
pandas
openpyxl
python-dotenv
assemblyai
langchain
langchain-community
langchain-google-genai
faiss-cpu
chromadb